Refactor catalog management out of pager
